--- nni/retiarii/strategies/__init__.py
@@ -1,2 +1,0 @@
-from .tpe_strategy import TPEStrategy
-from .random_strategy import RandomStrategy
