+++ test/retiarii_test/darts/test.py
@@ -4,8 +4,8 @@ import sys
 import torch
 from pathlib import Path
 
+from nni.retiarii import strategy
 from nni.retiarii.experiment import RetiariiExperiment, RetiariiExeConfig
-from nni.retiarii.strategies import TPEStrategy, RandomStrategy
 from nni.retiarii.trainer.pytorch import PyTorchImageClassificationTrainer
 
 from darts_model import CNN
@@ -18,10 +18,9 @@ if __name__ == '__main__':
                                                 optimizer_kwargs={"lr": 1e-3},
                                                 trainer_kwargs={"max_epochs": 1})
 
-    #simple_startegy = TPEStrategy()
-    simple_startegy = RandomStrategy()
+    simple_strategy = strategy.Random()
 
-    exp = RetiariiExperiment(base_model, trainer, [], simple_startegy)
+    exp = RetiariiExperiment(base_model, trainer, [], simple_strategy)
 
     exp_config = RetiariiExeConfig('local')
     exp_config.experiment_name = 'darts_search'
